+++ tests/test_ideal_gas.py
@@ -10,7 +10,7 @@ from pyturbo.thermo.ideal_gas import IdealDryAir, IdealGas
 class TestIdealGas:
     """Define tests for the ideal gas model."""
 
-    gas = IdealGas(r=287.058, cp=1004.0)
+    gas = IdealGas(287.058, 1004.0)
 
     t1 = 100.0
     t2 = 200.0
@@ -27,32 +27,32 @@ class TestIdealGas:
         assert self.cp == 1004.0
 
     def test_h(self):
-        assert self.gas.h(self.t1) == self.cp * self.t1
+        assert self.gas.h(self.t1) == pytest.approx(self.cp * self.t1, rel=1e-8)
 
     def test_gamma(self):
-        assert self.gas.gamma(self.t1) == self.cp / (self.cp - self.r)
+        assert self.gas.gamma(self.t1) == pytest.approx(self.cp / (self.cp - self.r), rel=1e-8)
 
     def test_phi(self):
-        assert self.gas.phi(self.t1) == self.cp * np.log(self.t1)
+        assert self.gas.phi(self.t1) == pytest.approx(self.cp * np.log(self.t1), rel=1e-8)
 
     def test_pr(self):
-        assert abs(self.pr - (self.t2 / self.t1) ** (self.eff * self.cp / self.r)) < 1e-6
-
-    def test_t_from_pr(self):
-        assert self.gas.t_from_pr(self.pr, self.t1, self.eff) == self.t2
-
-    def test_t_from_h(self):
-        assert self.gas.t_from_h(self.cp * self.t1) == self.t1
+        assert self.pr == pytest.approx(
+            (self.t2 / self.t1) ** (self.eff * self.cp / self.r), rel=1e-8
+        )
 
     def test_density(self):
-        assert self.gas.density(self.p1, self.t1) == self.p1 / (self.r * self.t1)
+        assert self.gas.density(self.p1, self.t1) == pytest.approx(
+            self.p1 / (self.r * self.t1), rel=1e-8
+        )
 
     def test_c(self):
-        assert self.gas.c(self.t1) == np.sqrt(self.gas.gamma(self.t1) * self.r * self.t1)
+        assert self.gas.c(self.t1) == pytest.approx(
+            np.sqrt(self.gas.gamma(self.t1) * self.r * self.t1), rel=1e-8
+        )
 
     def test_eff_poly(self):
-        assert (
-            abs(self.gas.eff_poly(self.p1, self.t1, self.p1 * self.pr, self.t2) - self.eff) < 1e-6
+        assert self.gas.eff_poly(self.p1, self.t1, self.p1 * self.pr, self.t2) == pytest.approx(
+            self.eff, rel=1e-8
         )
 
     def test_total_t(self):
@@ -73,9 +73,9 @@ class TestIdealGas:
         ts = 300.0
         tt = self.gas.total_t(ts, mach)
         pt = self.gas.total_p(ps, ts, tt)
-        assert ps == pytest.approx(self.gas.static_p(pt, tt, mach), 1e-3)
+        assert ps == pytest.approx(self.gas.static_p(pt, tt, mach, tol=1e-6), 1e-3)
 
-    def test_static_p_from_rhoV_subsonic(self):
+    def test_mach_f_wqa_subsonic(self):
         mach = 0.5
         ts = 300.0
         ps = 1e5
@@ -84,9 +84,9 @@ class TestIdealGas:
         pt = self.gas.total_p(ps, ts, tt)
         rhoV = self.gas.density(ps, ts) * mach * self.gas.c(ts)
 
-        assert ps == pytest.approx(self.gas.static_p_from_rhoV(pt, tt, rhoV, True), 1e-3)
+        assert mach == pytest.approx(self.gas.mach_f_wqa(pt, tt, rhoV, 1e-6, True), 1e-3)
 
-    def test_static_p_from_rhoV_supersonic(self):
+    def test_mach_f_wqa__supersonic(self):
         mach = 1.5
         ts = 300.0
         ps = 1e5
@@ -95,13 +95,13 @@ class TestIdealGas:
         pt = self.gas.total_p(ps, ts, tt)
         rhoV = self.gas.density(ps, ts) * mach * self.gas.c(ts)
 
-        assert ps == pytest.approx(self.gas.static_p_from_rhoV(pt, tt, rhoV, False), 1e-3)
+        assert mach == pytest.approx(self.gas.mach_f_wqa(pt, tt, rhoV, 1e-6, False), 1e-3)
 
 
 class TestDryAir:
     """Define tests for the dry air gas."""
 
-    gas = IdealGas(r=287.058, cp=1004.0)
+    gas = IdealGas(287.058, 1004.0)
     air = IdealDryAir()
 
     def test_cp(self):